move error middleware to app.ts
