+++ eslint-bridge/src/server.ts
@@ -20,10 +20,10 @@
 
 import 'module-alias/register';
 import http from 'http';
+
+import { createApp } from './app';
 import express from 'express';
 import { debug } from 'helpers';
-
-import createApp from './createApp';
 
 /**
  * Starts the bridge
